Increased performance w/ prepared insert stmt
After some CPU profiling w/ pprof, most of the time was being spent in
syscall, waiting on the database. This is the start of what will
probably be a small series of optimizations in an attempt to get to
1,110 URL/second write speeds.

This commit simply prepares the insert statement on the database when
the app is started to reduce the load on the DB to service what'll
ultimately be a very repeated query.

Tests:
- Requests serviced per second went from ~750 requests/second to ~850
  requests a second using `writeTest.js` and k6.