Negated the need to store the short url
After writing a function to decode URLs, I decided it would help us get
under our 10 year storage requirement if we had to persist less. Because
the ID is ultimately bijective (meaning it can be both encoded and
decoded both ways without any kind of data loss), we would only have to
save the ID in the database. Any app concerns can simply encode/decode
the ID on the fly before doing its work.

To maintain DB performance, I also changed the ordering of characters in
the base62 encoding scheme to be ascending for ASCII, as this
characteristic helps databases maintain more efficient indexes. I
rewrote the test answers to reflect this change.

Additionally, because shortening the URL requires asking the DB if the
url already exists, I also put that into a prepared statement that gets
created on the initialization of the app because preparing the insert
helped performance quite a bit.

Sadly, these changes do not improve the performance of writes. But given
that it does mean it's storing less data overall and getting closer to
our requirement, I'm committing it.

Tests:
- The test suite is green.
- The write test still remains able to write at ~850 urls/second.