Added read test: ~15,000 redirects/second.
This commit introduces a read test to the codebase and a small bugfix
to the app upon calls to redirect (the id array was being passed in
instead of a slice to the DB query formatting).

Four files were touched:

- `seed.py` was changed to created 8 workers to shorten urls.
  Additionally, it writes a JSON body to be used in the readTest.js
script.
- `readTest.js` defines the K6 read test. It makes sure that all
  responses for redirecting from a given short url return 200 and return
the expected long url.
- `gitignore` added the file name for the json file produced by
  `seed.py` since short urls will be different depending on the state of
the app.
- `repos.go` simply gets a slice view of the id array before passing in
  to the query that gets the redirect url (this happens in two places).

Note: This run hits our performance requirement of redirecting  11,600 a
second.

Tests:
- The test suite remains green.