Refactored and cleaned up project structure
After watching Mat Ryer's excellent GopherCon talk, "How I Write HTTP
Web Services after Eight Years" (and reviewing the corresponding
blogpost from February of this year w/ an update), I went about
refactoring a bit of `main` and the server setup. Previously, the routes
were being constructed as closures directly within `main` and had access
to the app as a consequence. However, I introduced a formal `server`
struct to contain all of logic that Gin needs and I augmented the
healthcheck to make sure it can ping the database.

On that note, I had previously handled DB connection is a fairly
erroneous way, though the result tended to be the same for Compose:
restart the application. I reworked the ping logic and added a member to
the database interface that allows it to be queried for connectivity.

Everything else is pretty much just organizing the project files,
refactoring names, and prettify-ing the whole lot. The idea was to use
more idiomatically correct Golang and make the project easier to parse
structurally from the root.

Tests:
- Test suite is still green
- Benchmarks still work as intended and aren't changed.